refactor: configuring routers for existing crews
